+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Support for running on obsolete Windows versions (useful for troubleshooting legacy systems).
 ### Changed
 * Increased the font size.
 * Re-wrote text handling to work both with and without Unicode support.